pyproject.toml
[project]
name = "crewai-a2a"
version = "0.1.0"
description = "CrewAI multi-agent crew with A2A protocol interop"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[a2a]>=1.14.4",
    "a2a-sdk>=1.0.0",
    "starlette>=0.40.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.5.0",
    "openai>=1.50.0",
]

[project.optional-dependencies]
graphiti = [
    "graphiti-core>=0.29.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# crewai[a2a] pins a2a-sdk ~0.3.x, but we need 1.0.x for the protobuf-based
# server stack. _a2a_compat.py bridges CrewAI's 0.3.x type imports.
override-dependencies = ["a2a-sdk>=1.0.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
